# Vectorize RMS norm variance using vectorize_read_with_alignment