@bbeckca 2025-10-04
@yewentao256  Thanks for your patience. In this PR, I'm only vectorizing RMS norm variance using vectorize_read_with_alignment. It seems in order to unlock more perf for vectorization on the write path, we'll need to update the util to support passing index used to access weights. Wonder if you could check if that sounds accurate?

https://github.com/vllm-project/vllm/blob/119f00630b8b6aea29b5c3141728f96963521df1/csrc/layernorm_kernels.cu#L39
https://github.com/vllm-project/vllm/blob/119f00630b8b6aea29b5c3141728f96963521df1/csrc/quantization/vectorization_utils.cuh#L46

cc @ProExpertProg