@ZJY0516 2025-10-05
> > Why do we have performance regression in many cases?
> 
> Thanks for the review! My understand is the slowdowns come from overhead from by calling vectorized path (indirection, packing/unpacking).
> 
> One way we could mitigate is by gating vectorized path to cases we see gains such as large tokens sequence and hidden size (T >= 1024 and H >= 8192). Further benchmarking can be done to find a sweet spot. Open to any feedback you may have on this.

I don't think it's a good idea. Relying on data from just one machine seems risky.