@bbeckca 2025-10-14
> @bbeckca Could you take a look at [buildkite/ci/pr/blackwell-test](https://buildkite.com/vllm/ci/builds/34875#0199e402-d2a1-472d-8dd8-baca508d603d)? Failure in
> 
> ```shell
> =========================================================================== FAILURES ===========================================================================
> --
>   | _________________________________________ test_all_reduce_fusion_pass_replace[dtype0-16-8-8-TestAllReduceRMSNormModel] _________________________________________
> ```
> 
> with pydantic validation error, perhaps related.
> 
> The version error in [quantization-test](https://buildkite.com/vllm/ci/builds/34875#0199e402-d27e-473e-9330-ae0e5fc2ba2d) should be fixed in main later

Sounds good. Took a quick look, this seems unrelated:
```
pydantic_core._pydantic_core.ValidationError: 1 validation error for ModelConfig
Value error, Invalid repository ID or local directory specified: 'nm-testing/TinyLlama-1.1B-Chat-v1.0-FP8-e2e'.
``